+++ src/components/panels/edit/modals/ComplexLookupModal.vue
@@ -1043,6 +1043,7 @@
                       Searching...
                     </option>
                     <template v-if="!isSimpleLookup()">
+                      <!-- .sort((a,b) => (a.label > b.label ? 1 : (a.label < b.label) ? -1 : 0)) -->
                       <option v-for="(r,idx) in activeComplexSearch.sort((a,b) => (a.label > b.label ? 1 : (a.label < b.label) ? -1 : 0))" :data-label="r.label" :value="r.uri" v-bind:key="idx" :style="(r.depreciated || r.undifferentiated) ? 'color:red' : ''" class="complex-lookup-result">
                         <div class="option-text">
                           <span v-html="generateLabel(r)"></span>